// FL410.Frontend/src/types/AeronaveViewModel.ts
export type AeronaveViewModel = {
    id: string;
    matricula: string;
    fabricante: string;
    anoFabricacao: string;
    modelo: string;
    numeroSerie: string;
    tipoICAO: string;
    categoriaHomologacao: string;
    tipoHabilitacao: string;
    classeAeronave: string;
    pmd: string;
    maxPAX: string;
    tipoVooAutorizado: string;
    tripMIN: string;
    numeroAssentos: string;
    categoriaRegistro: string;
    numeroMatricula: string;
    statusOperacao: string;
    gravame: string;
    validadeCVA: string;
    situacaoAeronavegabilidade: string;
}